Added some default config files for our  project
By default we connect to a test WiFi router so people should either edit the config file or replicate these settings
WDS mode proved to be incompatible with some devices so it was disabled.
AP mode used a bit too much RAM and also interfered with the local AP so it was also left disabled.